All known redmine status values are now processed.
Up to now, migration code only proceessed open tickets on redmine,
or tickets that have been open in the last migration run,
and since been closed.

This meant that a number of tickets
that had been closed prior to migration, or rejected,
or resolved, were not migrated.

Code updated such that all status values are migrated.